added build_uint()
moved the helper functions at the top to below the DO NOT TOUCH section (liable to removal potentially).

added a template function build_uint() that builds a larger uint from two uints of half that size.
overloaded for all double_int sizes, as well as std::size??_t (with the exception of 8-bit).

added a 2-parameter ctor to double_int to facilitate efficient build_uint.
double_int still functions identically to normal ints as long as this is not used.