bug fixes + optimization
fixed a bug where >>= for signed double_int types wasn't properly filling with sign bit (was still using trivial testing code from when signed variants were only an experiment).

made the conversion to u64 in << and >> explicit (will now work for any double_int type).

changed a few places that were taking their arguments by value instead of copying them.
initially, this was done for brevity, but with interlinking functions like this, it amounts to unnecessary copies.

fixed a few places where conversion between signed/unsigned was treated as implicit.

greatly optimized the ostream<< operator - now around 19 times faster for decimal.
the speedup is in extracting multiple digits (in this case 19 for decimal) as one block per divmod cycle and then processing them as per usual with a 64-bit built-in type, which is enormously faster.
this new algorithm is applied to hex and octal modes as well, but they don't have nearly the same speedup since they used bit shifting anyway (but this does have the benefit of performing large bit shifts all at once, which saves time proportional to the integer size).

added a function called low64() to get the low 64 bits of a value - used in the ostream<< operator to quickly extract a qword without using the cast operators, which are potentially (but probably not) slower than this method. (at any rate it can't hurt).