extended overloads
changed casts from reference to value - was by reference to avoid unnecessary copying, but may lead to dangling references.

reorganized the operator impl.
added a bunch of overloads to reduce the need for casting.
all ?= operators now have a generic version that takes a double_int& and a const T& - they all refer to the well-typed versions after a cast.

extended overloads

changed casts from reference to value - was by reference to avoid unnecessary copying, but may lead to dangling references.

reorganized the operator impl.
added a bunch of overloads to reduce the need for casting.
all ?= operators now have a generic version that takes a double_int& and a const T& - they all refer to the well-typed versions after a cast.