more overloads
added a whole bunch of overloads so that type casts aren't necessary for non-assignment binary operators.
to make it not ridiculous, they're all rolled into new macros that take the operator to extend.

did similar things to the comparison operators, which to avoid a stupid number of overloads, are rolled into a macro.