optimization
added the RESTRICT macro - currently unused.

removed the USE_C_PTR_TO_FIRST and USE_COMPACTNESS macros - they were getting annoying.
added a new macro called COMPACT(sign) that determines if that type of double_int is compact.
now, functions are SFINAE overloaded based on the result of a threshold value and the result of COMPACT(), as opposed to before when they were selected via #if (which could potentially cause issues if the assertion failed for the compact case.
this method entirely avoids the use of asserts.

added a bunch of optimized sfinae overloads - selected via new threshold macros on an operator-to-operator basis.
new overloads are showing substantial improvements, especially in decimal ostream<<, which is over 20x faster thanks to the addition of __divmod_unchecked and the compact overload for cmp().