major bug fixes
apparently at some point i commented out the vast majority of the test file and forgot to re-enable it.
this caused a ton of errors in the linux distro, which are now corrected for the most part.

fixed a bunch of ambiguous function calls resulting from using std::[u]int[dd]_t instead of the pure types.
the old way could (and was) potentially causing problems for unrepresented types.
i.e. unsigned long long is typically 64-bit, but is only typedefed as std::uint64_t in windows.
that means using unsigned long long would be an ambiguous call in any system that didn't uniquely typedef it as such.

fixed a bug where divmod with zero in the numerator and loop-based optimizations enabled failed.
this was caused by an off by one error that resulted in iterating over [0, 2^64) instead of [0, 0) due to an unsigned underflow.