string parsing
marked all double_int intrinsic functions as noexcept

added try_parse() and parse() as static methods of double_int.
this enables e.g. uint_t<256>::parse("123")
these both take std::string_view as the input, but are implemented via std::istringstream, which requires a string copy.

added unsigned overflow protection on parsing.
for decimal this is handled by performing the parsing on a double_int that it twice as large as the result and just guaranteeing the high half is always zero after merging a parsed block.
the alternative to this would be to use division logic (slow) or write special code specifically for this.

added signed overflow protection on parsing.
this just checks the sign bit and makes sure it's the same as the neg flag from the parsed input.

replaced the code in parsing decimal values that used (u64)pow(10, etc) with an integer we just multiply by 10 each step through.

added a bunch of test cases for string parsing.