reorganization and optimization
reorganized everything i don't want users to use to be in a nested 'detail' namespace.
for now only the uint_t and int_t template aliases are visible (will likely reintroduce some utilities like divmod later).

tidied up some of the code - some function overloads could be replaced with C++17 constexpr if.

major efficiency update:
apparently the cmp() function was being called a lot, and was using 54% of cpu time for the test program.
it was implemented in terms of < and >, which was causing more comparisons than were needed because < and > are defined in terms of a (recursively smaller) cmp() function.
cmp() is now implemented entirely in terms of itself and has dropped to only 11% cpu usage.