switched to loop-based logic
experimentally, i decided to switch double_int to use an array-based model instead of recursive halfing size ints.
the recursive way is fast for e.g. add/sub/bit arith, but not for e.g. shifts, which are pretty frequent.
it appears this new loop approach is a little over twice as fast, so after more testing this will be merged into master.