divmod() optimization
added a function named highest_set_bit() that gets the index of the highest set bit (or 0 if value is zero).
this is now used in divmod_unchecked() to calculate the upper bound bit of the value that needs to be inspected.
for typical integer values (which will typically not span the full integer range) this is substantially faster.
indeed, the test program now runs 11 times faster than it did before.
actually, it's so fast it's getting hard to measure (only about 40 ms).
thus, we will also need to add benchmarking code down the road to get a better feel for performance.

also, divmod() performs a lot of bit shifting.
i've rewritten the algorithm so that it does these in batches to reduce the total number of shift operations.
the way shift is written, it's effectively O(1) in bit shift count, so doing them in batches (e.g. <<= 4 instead of 4x <<= 1) is faster.
this is indeed faster for the test program, but that's likely due to e.g. testing 4096-bit integer stringification.
i've added a compile switch for selected which algorithm to use.
currently 256-bit and higher uses new algorithm, otherwise use the old (simpler) one to avoid the overhead.