interaction patches
fixed a few bugs that broke interactions between double_int and primitive types.
for instance, i had written the constructors/assignments for double_int to take uint64_t, uint32_t, etc...
but this doesn't cover built-in types which are equal size but not covered by those typedefs.
for instance int64 is long long and int32 is int, meaning long was not covered and would give ambiguity error (between int32, int64, and uint64).

cmp() now has overloads that take built-in types.
the whole suite of comparison operator overload are rewritten now.
they handle every case of built-in type comparison.
additionally cmp() with a built-in type is now optimized (no longer promotes the built-in to a (potentially very large) double_int type).

also, i've decided to flag cross-signage comparisons as a compile time error.
this means comparing unsigned double_int to signed anything is a compile error.
an issue arose regarding this when writing some new test cases and i decided it would be better to just not allow it at all.
(overload resolution would pick the double_int signage and implicitly cast the built-in type to the other signage without even issuing a warning).