added benchmarks + optimized multiply
added the first benchmark functions.
they're all the same function and use the same numbers, but the number of iterations and the operation to perform are controlled by params.

added benchmarks (using above) for divmod and multiply.

added a shift count optimization switch for multiply (works just like the one for divmod).
verified that the presets for the shift count switches are optimal (on my (windows) system, anyway).

since integer multiplication is commutative, i added extra logic to pick the fastest a * b vs. b * a.
this is now an impl function called _multiply() that takes pointer args to facilitate virtually-free arg swapping.
this is better than the alternative of calling it again with switched args because it avoids having to recompute the metrics that tell it if it should swap them in the first place (which are O(n)).
the metric used so far is whichever of them has the smaller highest set bit index (equivalently, which value is smaller).

the highest set bit of b arg is used as the upper bound for the loop, which avoids having to convert the value to bool (which is O(n)).
this also makes the 128-bit case faster, which is the most likely size for users to use.

the benchmarks don't assert that they got correct values.
that kind of thing would practically require the array of values and expected results to be passed as args.