added bigint type
added the bigint type, which is an arbitrary precision signed integer (2's complement).
i've nailed down the data contracts for the internal representation, but not much else is done.
it has a full suite of promotion constructors and assignment from built-in types and double_int.
it has a full suite of demotion conversions to built-in types.
eventually double_int will have a demotion constructor from bigint, but that's a much later concern.

implemented ++ and -- (pre and post) for bigint.
implemented == and != for comparing 2 bigints, and for comparing a bigint to a built-in type.
the comparison to a built-in type is O(1) in the number of blocks in the bigint due to the strict data contract for the internal representation - so that's neat.

simple test cases have been added for ++ and -- crossing the zero boundary.
tests with larger numbers will be added once + and - are implemented.
however, i suspect most of the useful test cases will only be added once the parsing and stringification functions are written.

added [[nodiscard]] to the postfix ++ and -- for both bigint and double_int.
for bigint this is because postfix results in a dynamic allocation (copy) and i want to slap the wrist of any user who might carelessly do this for efficiency reasons.
for double_int it's a similar issue, but not nearly as bad because the compiler is probably smart enough to avoid actually performing the copy due to it not being a dynamic allocation (more likely, anyway).