added bigint addition
added addition to bigint.
implemented in += for efficiency (does as little array modification as possible).
operator+ is overloaded to take any combination of rvalues/const lvalues.
the ones that take an rvalue move construct the temporary to avoid a copy.
the one that takes two rvalues orders them in such a way as to minimize the sum total amount of allocated space for both objects.