added bigint subtraction
i'd thought the subtraction logic was going to require adding the negative due to how taking the negative of a bigint can actually be larger.
e.g. 0x8000000000000000 (negative) converts to 0x00000000000000008000000000000000 when taking the negative (now positive).
however, after reconsidering, i determined the additional zeros cn just be thought of as part of the ghost blocks, in which case we just need to know what sign the value is expected to be.

thus, i've migrated the += bigint code to a template helper _add() which takes a bool flag as a template param.
if the flag is set, it will implicitly negate the second arg as it performs the addition.
thus subtraction has been added and is just as efficient as addition (no additional copies made).

i also added operator- with overloads for rvalues just like operator+ has.
the (const bigint&, bigint&&) overload actually performs -(b-a) which avoids a copy but will have to be benchmarked.
i assumed that it would be faster for very large bigint values and decided to roll with it.