bigint shifting and multiplication
added shifting and multiplication to the bigint type.
both of these have overloads of their non-compound forms to minimize bigint copying.
the multiplication has some extra work it has to do to guarantee negative numbers are multiplied correctly.
the actual multiplication is done by _multiply_positive().
_multiply_unknown() takes forwarding references to its arguments and then uses branch logic to decide what to do.
if a value is negative, it must be un-negated before the product, etc.
this design, however, means that nothing is copied/negated that doesn't need to be.
so e.g. multiplying two positive bigints doesn't trigger any copies/moves (aside from returning the result).

realized the signed multiply for double_int doesn't actually need to do its sign checking logic to perform the raw multiplication in 2's complement.
one of the nice properties of 2's complement is that signed and unsigned multiplication are equal up to the lower half of the full product.
despite the fact that the numbers are larger, we're still not computing the upper half of hte full product, so there's no need to handle it differently.
and if you did want the upper half, just case to a type twice as large, etc.
so the signed/unsigned versions are now the same template.