bigint comparisons and divmod
added comparison functions for bigint.
the == and != will still directly compare the block contents for efficiency.

new test cases for bigint comparisons lead to the discovery (and patch) of an incorrect piece of logic for comparing signed double_int.
the bug manifested when the left operand was negative and indeed less than the right.
the new (fixed) logic is actually cleaner and more streamlined - the best kind of bug fix.

added the divmod function for bigints.
added all its value type semantics overloads of operators as well.

with all of this, we have the groundwork necessary to create the serialization functions for operations on streams.
this will be in the next commit, along with more useful test cases for all the other operations (on larger values).
after that, the only missing operations for bigint will be and/or/xor.