finished basic impl for bigint
added and, or, and xor operations for bigint.
added istream and ostream integration for bigint.
added parsing and stringification tests for bigint (very basic for now - more later).
i was able to use templating to make the code that handles parsing/printing double_int and bigint to be the same code.
this will make it a lot more convenient if/when the impl hevavior needs to be changed (e.g. to handle the precision field).

caught a bug in the left shift operation for bigint where we weren't accounting for overflow out of the high block.
this is now fixed and everything is working properly.
a test case has been created that ensures this never breaks again.

due to the specific nature of left shifting a bigint, it happens that collapse is not needed if all we do is shift left by entire blocks.
thus, as a minor optimization, _collapse() is only called in the intra-block shift logic.
important note: this is not the case for right shifts.

changed the benchmark readouts to also show average time.
also switched them to microseconds because average times were frequently 0 milliseconds.