bug fix and pow() function
added the pow() function for bigint.
it uses a dynamic programming algorithm to minimize the number of multiplications that need to be performed.
by definition, i've made anything to the zero be 1 and anything to a negative power be 0.
this means that pow(0, 0) == 1, but i argue this is a more intuitive result than throwing an exception (not to mention being more performant).

this lead me to catch a bug in the compound assignment operators for division, modulus, and multiplication.
the overloads that took an lvalue and a const ref were moving from the lvalue.
this is potentially wrong, because the const ref arg could actually refer to the same variable, thus modifying it unintentionally.
this is fixed now by branching on whether or not they refer to the same object.
if they're different it means we can move from the lvalue, which avoids a dynamic allocation and an O(n) copy.

the logic used in the overload that takes 2 rvalues is ok as is (without branching) because it's already undefined behavior (canonically, not standardy) to move from the same value twice.