faster multiply algorithm
added a helper function that computes 64-bit products of unsigned values.
this has lots of little uses that'll be important later (for instance, the next point).

i was thinking about the multiply algorithm and decided it was stupid.
it causes lots of shifts and the time complexity is proportional to the number of set bits, not to mention it requires the weird highest set bit logic in order to be efficient, plus the swapping logic that only works because multiplication of integers commutes.
i've come up with a new algorithm based on the grade school method of long division (blast from the past).
basically, we compute the scaled product of one of the values by each 64-bit 'digit' of the other value and sum the results.
in benchmarking, this is nearly 10 times faster and represents the every case performance due to not being dependent on the number of set bits in either value.

i plan to extend this new algorithm to bigint.
additionally, this approach makes it possible to create overloads for handling double_int or bigint being multiplied by built-in types without having to up-convert them to double_int or bigint respectively.
this could be a significant speedup, especially for bigint due to avoiding a dynamic allocation.

did a bit more optimization stuff in various parts of the library, nothing nearly as fancy.