faster bigint multiply
rewrote the new double_int algorithm so that it would work with bigint.
the internal logic makes use of internal bigint values in invalid states (i.e. intermediate results that are not collapsed).
this can actually be extended quite a bit.
for instance, if the add operations are also allowed to be uncollapsed we can get away with having no intermadiate collapses on res and just do it once at the end of the function before returning.

this new algorithm gives around a 10-12 times speedup depending on the test case.

added a few more test cases for bigint multiply to make sure the new algorithm works.