gotta go fast
added a reduced, specialized implementation of addition in place of the adding step for the new double_int multiply algorithm.
for u128 this doesn't change much, but for u256 and u8196 (from benchmark) the elapsed times drop to zero.
i assume this is because, with this change, the optimizer somehow realizes what i'm doing and compiles into adc instructions.
added more test cases for double_int multiplication (same cases as some of the new bigint, but now also for u1024).