smaller bigint multiply logic
i realized all the temporary space was only used once and propagates unidirectionally.
this is indicative of not being necessary, which - in fact - it isn't.
i've rewritten the bigint multiply logic to be faster, smaller, and use no dynamically-allocated resources for book-keeping.
this is also a modest 10% faster.

also, changed the bigint multiply helper function to take references instead of pointers (no longer has need for arg swapping).