(probably) the last multiply speedup
applied the same new logic used to eliminate buffering in bigint multiplication to double_int.
this is less of an issue for double_int due to not using dynamic resources, but the algorithm itself is apparently also faster in and of itself.
this gives around a 40% speedup for the double_int benchmark cases.

also, i added volatile logic to the benchmarking functions to ensure they don't optimize to no-op (because after this commit it was back to doing that again).
and i also changed the double_int multiply helper to take references instead of pointers (like bigint now does) because we no longer need the arg swapping behavior that we used to rely on for speed.