msvc multiply intrinsics
added a new implementation of _mul_u64() called _mul_u64_fast().
this version is not constexpr but is allowed to use system-specific intrinsics.
added support for the msvc _umul128() intrinsic to make the computation faster.
this doesn't amount to much of a speedup for small values like u128, but doubles the performance of u8192.
thus, it gets more efficient the larger the value is, which makes it especially good for bigint.

the lack of constexpr is a bit annoying ... but constexpr isn't something that's used frequently.
maybe i'll just add a static utility that constexpr multiplies values if the need arises.
as it stands though, i'd much prefer to go for runtime speed than be a slave to constexpr limitations.