gcc multiply intrinsics
add intrinsic support for performing u64 multiplication using the gcc __uint128_t intrinsic.
ultimately this improves performance similarly to the msvc intrinsics in the windows release.
there was a weird thing that happened where the i8192 was actually slower than before (but u8192 was faster).
this was strange because they're calling essentially the same function (just different template instantiations of it, but on the same sizes).
apparently this was simply an ordering issue in the compiled executable, as moving the multiply benchmark before the divmod one normalizes the results.
i therefore assume this is caused by the signed template instantiation variant lying with an icache boundary inside of it, thus causing cache misses during execution.
the new ordering reflects the real efficiencies of the change (ordering issues are compiler-specific and not my problem).

conveniently, this makes multiply be the first benchmark to be run, and is indeed the faster of the two (so less annoying waiting for results this way anyway).