misc changes
fixed an issue with the intrinsics where it would no longer compile in 32-bit mode.
i strongly suggest (for speed) that users use 64-bit mode, but i'd also like it to work on 32-bit.

did some optimization work on divmod.
mostly this includes pre-allocating the result values, using unchecked bit tests, and avoiding unnecessary comparisons.
also added a non-negative bool < comparison function that theoretically makes the divmod comparisons faster.
experimental results show a negligible speedup of (~4%).
these changed are applied to both double_int and bigint.

fixed a bug where printing vales wasn't behaving identically to the standard C++ ostream<< behavior.
this is now fixed and is actually much cleaner as far as the wrappers are concerned.
moved hex ostream<< to be the first base flag checked (in case multiple options are set for some reason - this should never happen).
changed the ostream<< printing to reverse the string and print it rather than print character by character in reverse order.
apparently this is actually significantly faster (nearly 2x speedup for hex/oct - less so for dec but still faster).
ostream<<(string) internally is implemented like this, but directly writes to rdbuf() and is apparently significantly faster.

added the bigint::factorial() static function.
this takes u64 instead of bigint because (2^64)! would take more memory to store than i anticipate a computer will ever have.
the first several (as many as possible) values of n! are pre-computed as u64 in a lookup table.

added benchmark thingys for printing.