bigint printing changes
the last commit introduced an infinite loop when printing a negative bigint value in hex or oct.
this is because those forms use (arithmetic) right shifts and will therefore never reach 0 if the value is negative (stuck at -1).
the fix is to add a leading 0 block if it's negative before doing the rest of the logic.

additionally, changed the printing behavior of bigint when printing in hex/oct mode.
the new rule for printing bigint is: the smallest string such that sign extension yields the same value.
this means -1 printed in hex yields "f", and in oct yields "7" instead of "177...77".
this also means 15 printed in hex yields "0f", and 7 in oct yields "07".
as you can see, chars may be modified/appended to the front of the string to ensure the value is preserved after sign extension

the reason this had to be changed is because parsing "ffffffffffffffff" as hex would be ambiguous (either u64 max or -1).
as a consequence, conversions from bigint-encoded strings to non-bigint types via parsing is now broken.
i'll just assume that if you write something as a bigint you'll also read it as a bigint.
the same goes for the other types.
if a different type is needed, parse as the type used to write it, then cast the value to the required type.

another oddity (for oct) is sometimes two '0' chars are printed preceding the "real" value.
e.g. 7 in oct yields "07" ordinarily, or "007" if showbase is set (first 0 is octal prefix, second is to ensure proper sign extension).

the parsing functions to go along with these have not yet been updated, and therefore will not work.
decimal, however, will still work and should really be the only desired format, as it transcends the differences in encodings of the various integral types.
parsing will be fixed in the next commit.