beginning of massive refactoring
so even with the simple test file the project was taking like 16 seconds to compile.
it's obviously due to the ridiculous number of large templates that are being unrolled for the fixed_int type (previously named double_int).
so i've decided to rewrite all the fixed_int stuff in terms of an antipolymorphic wrapper.
for small number of bits, this will be slower (larger should be the same).
 however, the real benefit is that it will make compile times significantly faster.
it also gives an opportunity to define functions in different files.
the project is no longer header-only due to obvious limitations.

this commit refactors everything but divmod into the new system.
this includes all other operators, as well as tostring and parse.
i've also managed to eliminate a lot of the uses of preprocessing for function declaration.
the basic pattern is that any template function (i.e. fixed_int function) just calls an antipolymorphic helper function).