bug fix and more cleanup
fixed a bug where we ere using the wrong signedness logic for remainder of division.
apparently since ISO2011 % is guaranteed to return the same sign as the numerator.
we were returning the "same" sign that the quotient would be (not counting rounding to zero).

continued on the refactor ad cleaned up a bunch of stuff.
some more template functions were rewritten with fixed_int_wrapper helpers to cut down on template function sizes.

completed the full public interface for detail::cmp() for all biggerints types.
this is then used to greatly simplify the comparison operators logic: everything just calls cmp().
unless i can think of a reason not to, i intend to export cmp() into the core biggerints namespace for users to access.