added `tostr()` function
added the `tostr()` function for converting values to strings.
this allows users to directly convert to string instead of having to go through `ostringstream`.
this results in a 30% speedup for decimal, a 400% speedup for hex, and a 350% speedup for oct.
so this method is significantly faster if you just want a string.
printing values is still performed by converting to string and printing the result.

tostr format settings are now controlled by a struct named `tostr_fmt`.
this also has an overloaded `operator()` for performing the conversion.
the intent is to have tostr_fmt be considered a formatter (i.e. a function), not just a collection of settings.
the `tostr()` function just calls `operator()` on the object.
this mostly exists for the case when the user just wants a default constructed formatter:
i.e. `tostr(val)` instead of `tostr_fmt{}(val)`

tostr logic is now moved outside of the detail namespace so as to be visible without adl.
did a bit more cleanup and reorganizing.
all the public stuff (stuff for users) will be compressed in one section.
operators will remain in detail namespace and just use adl lookup so as not to clutter the exterior namespace.