tostr/parse updates
added base 2 as a valid tostr/parse base.

made tostr logic faster by adding alloc reserve logic for the generated string.
for every base except decimal this is significantly faster (~20-40%).
decimal's time is taken up mostly by divmod, so that's the bottleneck for it.

changed parse_fmt's base constructor to implicit so that e.g. val.tostr(16) is valid.
otherwise this would have to be val.tostr(tostr_fmt{16}) or tostr_fmt{16}(val).

added back the parse_fmt type.
it has a new collection of functions.
takes a dest and a string to act as try_parse().
takes just a string and acts as parse().
takes an istream and acts as extract().

changed octal to use the "0o" prefix instead of C-style "0", which is confusing.

fixed a bug in octal parser where we had "false;" instead of "return false;".
new test cases handle this.