mixed type operators
the standard behavior of most operators in C++ allows more implicit conversions than we used to support.
this includes a lot of extra overloads of everything, notably compound assignment.
i took the oportunity to add efficient template versions of them.
these are added for +=, -=, &=, |=, ^=, *=, /=, and %=.
the non compound-assignment operators also have been rewritten, but mul/div aren't finished yet.

added a bunch of test cases for ensuring all of these work.
most of the tests are compile time tests for the return types of non compound-assignment ops.
the compound assignments have new test cases for all the implicit conversion behaviors.

the new impl for compound assignment stuff enables the use of slicing instead of truncation for large types on the right hand side.