unrestricted cmp + bug fix
removed the restriction that signed/unsigned values cannot be compared.
the idea was based on the fact that C++ itself will emit warnings in these cases, but it was getting annoying.
the defined behavior in this case follows intuitively: the comparison is performed *as if* both values were cast to bigint.

optimized some of the inefficiencies in the * and *= for fixed_int.
*= now enables slicing for the scale factor to avoid a copy/cast when it can.
this introduced a self-referencing bug that had to be fixed.
the other compound assignment ops will be tested more thoroughly in the next commit for this same type of bug.

when unary '-' was being applied to unsigned fixed_int it was actually upcasting to bigint and using that one because unsigned has no unary-.
an overload of unary operator- is now added for unsigned types and is explicitly deleted.

removed unnecessary symmetry overloads of some comparison stuff.
apparent symmetry is now performed by cmp_helper().
old things that used to be called cmp_helper() are now called cmp_raw() to reflect lack of symmetry.

fixed a bug where signed comparisons of fixed_int to a (signed) builtin type where the fixed_int was truncatable to 1 block was getting the opposite of expected result.
this was an old bug from the wrapped/wrapped comparison code that was fixed there but not here.
corrected code is invariant of signedness.

old comparability sfinae only required one of the types be a biggerints type.
this allowed e.g. bigint{} == "hello world" to *appear* to work at the call site but fail to compile.
new comparability asserts that one must be a biggerints type and the other must be either a biggerints type or a builtin integral type.